fix: preserve enforced model specs through agents parsing
